requirements.txt
Django==4.0.5
psycopg2-binary==2.9.3
gunicorn==20.1.0
Pillow==9.1.1
django-mptt==0.13.4
ipython==8.4.0
martor==1.6.13
python-dateutil==2.8.2
django-markdownx==4.0.0-beta.1
